Gem require and misc fixes
